feat: Implement session chat

+++ memory/api/memory_api/routers/models/protocol.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 class Model(BaseModel):
     model_name: str
     max_length: int
-    updated_at: int
+    updated_at: float
     default_settings: dict
